+++ packages/react-native-audio-api/common/cpp/utils/AudioUtils.h
@@ -6,7 +6,7 @@
 
 namespace audioapi::AudioUtils {
 
-size_t timeToSampleFrame(double time, int sampleRate);
+size_t timeToSampleFrame(double time, float sampleRate);
 double sampleFrameToTime(int sampleFrame, int sampleRate);
 
 float linearInterpolate(const float *source, size_t firstIndex, size_t secondIndex, float factor);
